requirements.txt
pywebview>=5.2
